SYNC-1279 Files can become corrupt if a version is deleted

+++ modules/util/sync-engine/src/com/liferay/sync/engine/documentlibrary/handler/DownloadFileHandler.java
@@ -15,6 +15,7 @@
 package com.liferay.sync.engine.documentlibrary.handler;
 
 import com.liferay.sync.engine.documentlibrary.event.Event;
+import com.liferay.sync.engine.documentlibrary.util.FileEventUtil;
 import com.liferay.sync.engine.filesystem.Watcher;
 import com.liferay.sync.engine.filesystem.util.WatcherRegistry;
 import com.liferay.sync.engine.model.SyncAccount;
@@ -87,7 +88,32 @@ public class DownloadFileHandler extends BaseHandler {
 
 		SyncFile syncFile = getLocalSyncFile();
 
-		SyncFileService.deleteSyncFile(syncFile, false);
+		if ((Boolean)getParameterValue("patch")) {
+			FileEventUtil.downloadFile(getSyncAccountId(), syncFile);
+		}
+		else {
+			SyncFileService.deleteSyncFile(syncFile, false);
+		}
+	}
+
+	@Override
+	public boolean handlePortalException(String exception) throws Exception {
+		if (exception.equals(
+				"com.liferay.portlet.documentlibrary." +
+					"NoSuchFileVersionException") &&
+			(Boolean)getParameterValue("patch")) {
+
+			if (_logger.isDebugEnabled()) {
+				_logger.debug("Handling exception {}", exception);
+			}
+
+			FileEventUtil.downloadFile(
+				getSyncAccountId(), (SyncFile)getParameterValue("syncFile"));
+
+			return true;
+		}
+
+		return super.handlePortalException(exception);
 	}
 
 	protected void copyFile(
